[v3] Moves JsonMapper into line with the v3 structure
