+++ tests/Unit/JsonMapperTest.php
@@ -3,8 +3,8 @@
 namespace Unit;
 
 use PHPAlchemist\Exception\BadJsonException;
-use PHPAlchemist\Services\JsonMapper;
-use PHPAlchemist\Traits\JsonHydratorTrait;
+use PHPAlchemist\Service\JsonMapper;
+use PHPAlchemist\Trait\JsonHydratorTrait;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 
@@ -18,7 +18,7 @@ trait FizSetter
 
 abstract class AbstractHydratorClass
 {
-    public ?string $foo    = null;
+    public ?string $foo = null;
     protected ?string $bar = null;
     protected ?string $fiz = null;
 
@@ -53,14 +53,15 @@ class MockJsonBadHydrationClass extends AbstractHydratorClass
     private ?string $buzz = null;
 
     // buzz
-    public function setBuzz(string $buzz) : void
-    {
-        $this->buzz = $buzz;
-    }
 
     public function getBuzz() : ?string
     {
         return $this->buzz;
+    }
+
+    public function setBuzz(string $buzz) : void
+    {
+        $this->buzz = $buzz;
     }
 }
